testimpl

+++ src/com/meituan/dao/BusiDAO.java
@@ -6,9 +6,9 @@ import com.meituan.domain.Busi;
 public interface BusiDAO
 {
 	//查询账号密码
-	public int getCountForBusi(String busiPhone,String busiPass);
+	public long getCountForBusi(String busiPhone,String busiPass);
 	//查询电话号码
-	public int getCountForPhone(String busiPhone);
+	public long getCountForPhone(String busiPhone);
 	//保存注册信息
 	public void sava(Busi busi);
 }
