+++ src/com/meituan/dao/UserDAO.java
@@ -6,9 +6,9 @@ import com.meituan.domain.User;
 public interface UserDAO
 {
 	//查询账号密码
-	public int getCountForUser(String userPhone,String userPass);
+	public long getCountForUser(String userPhone,String userPass);
 	//查询电话号码
-	public int getCountForPhone(String userPhone);
+	public long getCountForPhone(String userPhone);
 	//保存注册信息
 	public void sava(User user);
 }
